Move Widgets into Widget Folder Move Data Window and associated classes to Windows.Data subnamespace
