+++ Dalamud/Interface/Internal/Windows/Data/Widgets/AddonInspectorWidget.cs
@@ -1,4 +1,4 @@
-﻿namespace Dalamud.Interface.Internal.Windows;
+﻿namespace Dalamud.Interface.Internal.Windows.Data;
 
 /// <summary>
 /// Widget for displaying addon inspector.
