+++ Dalamud/Interface/Internal/Windows/Data/Widgets/AddonWidget.cs
@@ -3,7 +3,7 @@ using Dalamud.Memory;
 using Dalamud.Utility;
 using ImGuiNET;
 
-namespace Dalamud.Interface.Internal.Windows;
+namespace Dalamud.Interface.Internal.Windows.Data;
 
 /// <summary>
 /// Widget for displaying Addon Data.
