+++ Dalamud/Interface/Internal/Windows/Data/Widgets/ConditionWidget.cs
@@ -1,7 +1,7 @@
 ﻿using Dalamud.Game.ClientState.Conditions;
 using ImGuiNET;
 
-namespace Dalamud.Interface.Internal.Windows;
+namespace Dalamud.Interface.Internal.Windows.Data;
 
 /// <summary>
 /// Widget for displaying current character condition flags.
