+++ Dalamud/Interface/Internal/Windows/Data/Widgets/DtrBarWidget.cs
@@ -2,7 +2,7 @@
 using Dalamud.Game.Gui.Dtr;
 using ImGuiNET;
 
-namespace Dalamud.Interface.Internal.Windows;
+namespace Dalamud.Interface.Internal.Windows.Data;
 
 /// <summary>
 /// Widget for displaying dtr test.
