+++ Dalamud/Interface/Internal/Windows/Data/Widgets/FateTableWidget.cs
@@ -1,7 +1,7 @@
 ﻿using Dalamud.Game.ClientState.Fates;
 using ImGuiNET;
 
-namespace Dalamud.Interface.Internal.Windows;
+namespace Dalamud.Interface.Internal.Windows.Data;
 
 /// <summary>
 /// Widget for displaying the Fate Table.
