+++ Dalamud/Interface/Internal/Windows/Data/Widgets/GamepadWidget.cs
@@ -3,7 +3,7 @@
 using Dalamud.Game.ClientState.GamePad;
 using ImGuiNET;
 
-namespace Dalamud.Interface.Internal.Windows;
+namespace Dalamud.Interface.Internal.Windows.Data;
 
 /// <summary>
 /// Widget for displaying gamepad info.
