+++ Dalamud/Interface/Internal/Windows/Data/Widgets/ImGuiWidget.cs
@@ -4,7 +4,7 @@ using Dalamud.Interface.Internal.Notifications;
 using Dalamud.Interface.Windowing;
 using ImGuiNET;
 
-namespace Dalamud.Interface.Internal.Windows;
+namespace Dalamud.Interface.Internal.Windows.Data;
 
 /// <summary>
 /// Widget for displaying ImGui test.
