+++ Dalamud/Interface/Internal/Windows/Data/Widgets/ObjectTableWidget.cs
@@ -7,7 +7,7 @@ using Dalamud.Game.Gui;
 using Dalamud.Utility;
 using ImGuiNET;
 
-namespace Dalamud.Interface.Internal.Windows;
+namespace Dalamud.Interface.Internal.Windows.Data;
 
 /// <summary>
 /// Widget to display the Object Table.
