+++ Dalamud/Interface/Internal/Windows/Data/Widgets/ServerOpcodeWidget.cs
@@ -2,7 +2,7 @@
 using ImGuiNET;
 using Newtonsoft.Json;
 
-namespace Dalamud.Interface.Internal.Windows;
+namespace Dalamud.Interface.Internal.Windows.Data;
 
 /// <summary>
 /// Widget to display the currently set server opcodes.
