+++ Dalamud/Interface/Internal/Windows/Data/Widgets/TargetWidget.cs
@@ -3,7 +3,7 @@ using Dalamud.Game.ClientState.Objects;
 using Dalamud.Utility;
 using ImGuiNET;
 
-namespace Dalamud.Interface.Internal.Windows;
+namespace Dalamud.Interface.Internal.Windows.Data;
 
 /// <summary>
 /// Widget for displaying target info.
