+++ Dalamud/Interface/Internal/Windows/Data/Widgets/ToastWidget.cs
@@ -3,7 +3,7 @@
 using Dalamud.Game.Gui.Toast;
 using ImGuiNET;
 
-namespace Dalamud.Interface.Internal.Windows;
+namespace Dalamud.Interface.Internal.Windows.Data;
 
 /// <summary>
 /// Widget for displaying toast test.
